+++ timepiece/src/main/java/com/appcenter/timepiece/repository/NotificationRepository.java
@@ -1,5 +1,6 @@
-package com.appcenter.timepiece.notify;
+package com.appcenter.timepiece.repository;
 
+import com.appcenter.timepiece.domain.Notification;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
